requirements.txt
# =============================================================================
# requirements.txt - geek.bidu.guru
# Dependencias Python do projeto
# =============================================================================

# -----------------------------------------------------------------------------
# Framework Web
# -----------------------------------------------------------------------------
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# -----------------------------------------------------------------------------
# Templates (SSR)
# -----------------------------------------------------------------------------
jinja2==3.1.4

# -----------------------------------------------------------------------------
# Banco de Dados
# -----------------------------------------------------------------------------
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# -----------------------------------------------------------------------------
# Cache
# -----------------------------------------------------------------------------
redis==5.2.1

# -----------------------------------------------------------------------------
# Validacao e Configuracao
# -----------------------------------------------------------------------------
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# -----------------------------------------------------------------------------
# Seguranca
# -----------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# -----------------------------------------------------------------------------
# HTTP Client (para APIs externas)
# -----------------------------------------------------------------------------
httpx==0.28.1
aiohttp==3.11.10

# -----------------------------------------------------------------------------
# Utilitarios
# -----------------------------------------------------------------------------
python-slugify==8.0.4
python-dateutil==2.9.0.post0
pytz==2024.2

# -----------------------------------------------------------------------------
# Processamento de Imagens
# -----------------------------------------------------------------------------
pillow==11.0.0

# -----------------------------------------------------------------------------
# Logging
# -----------------------------------------------------------------------------
python-json-logger==2.0.7
structlog==24.4.0

# -----------------------------------------------------------------------------
# Rate Limiting
# -----------------------------------------------------------------------------
slowapi==0.1.9

# -----------------------------------------------------------------------------
# Markdown (para conteudo)
# -----------------------------------------------------------------------------
markdown==3.7
bleach==6.2.0

# -----------------------------------------------------------------------------
# LLM / IA (Geracao de Conteudo)
# -----------------------------------------------------------------------------
litellm==1.54.1
openai==1.58.1

# -----------------------------------------------------------------------------
# SEO / Sitemap
# -----------------------------------------------------------------------------
# (geracao manual via FastAPI)

# -----------------------------------------------------------------------------
# Desenvolvimento e Testes
# -----------------------------------------------------------------------------
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
aiosqlite==0.20.0
black==24.10.0
ruff==0.8.2
mypy==1.13.0
